+++ scripts/module.js
@@ -9,7 +9,7 @@ import { importItems, loadItemDatabases } from "./importers/items.js"
 
 const CHARACTER_TYPE_MAP = {
     0: 'Character',
-    1: 'NPC (BETA - data corruption may result)',
+    1: 'NPC',
 }
 
 Hooks.once('ready', async function() {
@@ -32,12 +32,27 @@ function getActorSheetHeaderButtons(sheet, buttons) {
 
 Hooks.on('getActorSheetHeaderButtons', getActorSheetHeaderButtons)
 
+function isUsingMookSheet(actor) {
+    console.log(actor.id, actor.flags.core);
+    const currentSheetClass = actor?.flags?.core?.sheetClass ||
+        game.settings.get("core", "sheetClasses")?.Actor?.mook;
+    return currentSheetClass === 'cyberpunk-red-core.CPRMookActorSheet';
+}
+
+let currentDialog;
+
 function startImport(sheet) {
-    new Dialog({
+    if (currentDialog !== undefined) {
+        currentDialog.close();
+    }
+    const newDialog = new Dialog({
         title: 'Import Character from cyberpunkred.com',
         content: `Enter Character Export Code:
     <input class="character-import-code"><br>
-    <div class="character-import-name">&nbsp;</div>`,
+    <div class="character-import-text">
+      <div class="character-import-name">&nbsp;</div>
+      <div class="character-import-message">&nbsp;</div>
+    </div>`,
         buttons: {
             import: {
                 icon:  '<i class="fas fa-cloud-download-alt"></i>',
@@ -48,6 +63,11 @@ function startImport(sheet) {
                     await importCharacter(data, sheet.object)
                 }
             }
+        },
+        close: html => {
+          if (currentDialog === newDialog) {
+              currentDialog = undefined;
+          }
         },
         render: html => {
             const button = html.find('button');
@@ -68,6 +88,14 @@ function startImport(sheet) {
                         button.prop('disabled', false);
                         const characterType = CHARACTER_TYPE_MAP[characterData.character_type_id];
                         nameDisplay.text(`${characterType} to Import: ${characterData.name}`);
+                        if (isUsingMookSheet(sheet.object)) {
+                            html.find('.character-import-message').text(
+                                'This actor is currently using the Mook sheet. In order to import' +
+                                ' successfully, this actor will be temporarily set to use the player' +
+                                ' character sheet during the import process then restored at the end.')
+                        } else {
+                            html.find('.character-import-message').html('&nbsp;');
+                        }
                     } catch (e) {
                         nameDisplay.text(e);
                         nameDisplay.addClass('invalid-code');
@@ -84,7 +112,9 @@ function startImport(sheet) {
                 }
             })
         }
-    }).render(true);
+    });
+    currentDialog = newDialog;
+    newDialog.render(true);
 }
 
 async function importCharacter(data, actor) {
@@ -92,6 +122,13 @@ async function importCharacter(data, actor) {
         throw new Error('Can only import to characters and mooks');
     }
     console.info('Importing character', data, 'to actor', actor);
+
+    const originalSheetClass = actor?.flags?.core?.sheetClass ?? '';
+    const mustReconfigureSheetClass = isUsingMookSheet(actor);
+    if (mustReconfigureSheetClass) {
+        console.warn(`Temporarily configuring ${actor} to use the Player Character sheet during import.`);
+        await actor.update({ flags: { core: { sheetClass: 'cyberpunk-red-core.CPRCharacterActorSheet' } } });
+    }
 
     const forWhom = `${data.name} from ${data.code_to_character}`;
 
@@ -107,6 +144,10 @@ async function importCharacter(data, actor) {
         const errorMessage = `Failed to import ${forWhom}.`;
         ui.notifications.error(errorMessage);
         console.error(errorMessage, e);
+    }
+
+    if (mustReconfigureSheetClass) {
+        await actor.update({ flags: { core: { sheetClass: originalSheetClass } } });
     }
 
     ui.notifications.info(`Done importing character ${forWhom}. Max Humanity and Empathy may need